requirements.txt
click==8.1.7
exceptiongroup==1.2.1
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
Pint==0.23
pluggy==1.5.0
Pygments==2.18.0
pytest==8.2.0
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
xmltodict==0.13.0
